Support wide M4A durations and offsets with bounded streaming

+++ src/main/file/m4a/M4aAlacReader.hpp
@@ -17,7 +17,7 @@ namespace cupuacu::file::m4a
         std::uint32_t sampleRate = 0;
         std::uint16_t channels = 0;
         std::uint16_t bitDepth = 0;
-        std::uint32_t frameCount = 0;
+        std::uint64_t frameCount = 0;
         cupuacu::SampleFormat sampleFormat = cupuacu::SampleFormat::Unknown;
         std::vector<std::uint8_t> interleavedPcmBytes;
         std::vector<cupuacu::DocumentMarker> markers;
@@ -37,7 +37,7 @@ namespace cupuacu::file::m4a
         std::uint32_t sampleRate = 0;
         std::uint16_t channels = 0;
         std::uint16_t bitDepth = 0;
-        std::uint32_t frameCount = 0;
+        std::uint64_t frameCount = 0;
         cupuacu::SampleFormat sampleFormat = cupuacu::SampleFormat::Unknown;
         std::vector<cupuacu::DocumentMarker> markers;
     };
